+++ examples/threaditjs/app.js
@@ -80,10 +80,10 @@ var Header = {
 			m("p.head_links", [
 				m("a[href='https://github.com/koglerjs/threaditjs/tree/master/examples/mithril']", "Source"),
 				" | ",
-				m("a[href='http://threaditjs.com']", "ThreaditJS Home"),
+				m("a[href='https://threaditjs.com']", "ThreaditJS Home"),
 			]),
 			m("h2", [
-				m("a[href='/']", {oncreate: m.route.link}, "ThreaditJS: Mithril"),
+				m(m.route.Link, {href: "/"}, "ThreaditJS: Mithril"),
 			]),
 		]
 	}
@@ -102,7 +102,7 @@ var Home = {
 					threads.map(function(thread) {
 						return [
 							m("p", [
-								m("a", {href: "/thread/" + thread.id, oncreate: m.route.link}, m.trust(T.trimTitle(thread.text))),
+								m(m.route.Link, {href: "/thread/" + thread.id}, m.trust(T.trimTitle(thread.text))),
 							]),
 							m("p.comment_count", thread.comment_count + " comment(s)"),
 							m("hr"),