+++ render/hyperscript.js
@@ -93,7 +93,7 @@ function hyperscript(selector) {
 		vnode.children = Vnode.normalizeChildren(vnode.children)
 		if (selector !== "[") return execSelector(selectorCache[selector] || compileSelector(selector), vnode)
 	}
-	
+
 	vnode.tag = selector
 	return vnode
 }