+++ render/tests/test-updateText.js
@@ -9,7 +9,7 @@ o.spec("updateText", function() {
 	o.beforeEach(function() {
 		$window = domMock()
 		root = $window.document.createElement("div")
-		render = vdom($window).render
+		render = vdom($window)
 	})
 
 	o("updates to string", function() {