+++ scripts/.eslintrc.js
@@ -1,0 +1,9 @@
+module.exports = {
+    "extends": "../.eslintrc.js",
+	"env": {
+        "browser": null,
+	},
+    "parserOptions": {
+        "ecmaVersion": 2019,
+    },
+};